+++ cds-plugin.js
@@ -9,8 +9,7 @@ cds.env.cov2ap.build = cds.env.cov2ap.build === undefined ? true : cds.env.cov2a
 
 if (cds.env.cov2ap.plugin) {
   cds.on("bootstrap", async (app) => {
-    app.use(cov2ap());
-    cov2ap._pluginActive = true;
+    app.use(cov2ap.singleton());
   });
   if (cds.env.cov2ap.build && cds.build && cds.build.register) {
     cds.build.register("cov2ap", {