+++ src/build.js
@@ -9,7 +9,11 @@ module.exports = class COV2APBuildPlugin extends cds.build.BuildPlugin {
     cds.env.cov2ap = cds.env.cov2ap || {};
     cds.env.cov2ap.plugin = cds.env.cov2ap.plugin === undefined ? true : cds.env.cov2ap.plugin;
     cds.env.cov2ap.build = cds.env.cov2ap.build === undefined ? true : cds.env.cov2ap.build;
-    return cds.env.cov2ap && cds.env.cov2ap.plugin;
+    return (
+      cds.env.cov2ap.plugin &&
+      cds.env.cov2ap.build &&
+      (cds.env.profiles?.includes("with-mtx-sidecar") || !!cds.env.requires["cds.xt.ModelProviderService"])
+    );
   }
 
   init() {