+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 1.12.2 - 2024-02-xx
+## Version 1.12.2 - 2024-01-09
 
 ### Fixed
+
+- Build task only applicable if MTX sidecar is active
+- Use adapter singleton in CDS plugin
 
 ## Version 1.12.1 - 2024-01-08